fix: passkey bech32
