+++ packages/sdk/src/modules/index.ts
@@ -3,4 +3,3 @@ export * from './coders';
 export * from './service';
 export * from './provider';
 export * from './webauthn';
-export * from './passkey';
