--- packages/sdk/src/modules/passkey/index.ts
@@ -1,2 +1,0 @@
-export * from './address';
-export * from './types';
