--- packages/sdk/src/modules/passkey/types.ts
@@ -1,5 +1,0 @@
-export enum Batch32Prefix {
-  PASSKEY = 'passkey',
-}
-
-export type Batch32 = `${Batch32Prefix}.${string}`;
